self made battery box
Hello,
To enhance the cost-effectiveness of the otoscope, I have developed a printable version of the battery box instead of the recommended purchase.

It maintains the original size of the recommended one, ensuring a perfect fit in the "Battery holder".

To further improve cost-effectiveness and reduce dependency on online material ordering, I have created two versions of electrical connections. One utilizes a self-made spring from any conducting metal wire, soldered to the wires. The other version uses parts purchased on Amazon for less than a dollar per set.