Specula holder adaptation to standard
Hello

I made a new version of the specula holder to make it possible to use with the standard disposable specula.

This kind of speculum is made of smooth plastic that makes it easily washable for sanitary purpose. This model is the most common, with two possible sizes.

The piece should be mounted with the hole up.